+++ packages/designer/src/simulator.ts
@@ -142,7 +142,7 @@ export interface ISimulatorHost<P = object> extends ISensor {
   computeComponentInstanceRect(instance: ComponentInstance, selector?: string): DOMRect | null;
 
   findDOMNodes(instance: ComponentInstance, selector?: string): Array<Element | Text> | null;
-
+  
   /**
    * 销毁
    */