+++ packages/editor-preset-vision/src/pages.ts
@@ -1,6 +1,7 @@
 import { designer } from './editor';
 import { RootSchema } from '@ali/lowcode-types';
 import { DocumentModel } from '@ali/lowcode-designer';
+import NodeCacheVisitor from './rootNodeVisitor';
 
 const { project } = designer;
 
@@ -38,10 +39,13 @@ const pages = Object.assign(project, {
     } else {
       componentsTree = pages[0].componentsTree;
       if (componentsTree[0]) {
-        componentsTree[0].componentName = 'Page';
+        componentsTree[0].componentName = componentsTree[0].componentName || 'Page';
         // FIXME
-        componentsTree[0].lifeCycles = {};
-        componentsTree[0].methods = {};
+        if (componentsTree[0].componentName === 'Page' ||
+        componentsTree[0].componentName === 'Component') {
+          componentsTree[0].lifeCycles = {};
+          componentsTree[0].methods = {};
+        }
       }
     }
 
@@ -95,5 +99,16 @@ Object.defineProperty(pages, 'currentPage', {
     return project.currentDocument;
   }
 })
+
+pages.onCurrentPageChange((page: DocumentModel) => {
+  if (!page) { return; }
+  page.acceptRootNodeVisitor('NodeCache', (rootNode) => {
+    const visitor: NodeCacheVisitor = page.getRootNodeVisitor('NodeCache');
+    if (visitor) {
+      visitor.destroy();
+    }
+    return new NodeCacheVisitor(page, rootNode);
+  });
+});
 
 export default pages;