+++ packages/react-renderer/src/engine/index.jsx
@@ -23,6 +23,7 @@ const ENGINE_COMPS = {
   BlockEngine,
   AddonEngine,
   TempEngine,
+  // DivEngine: BlockEngine,
 };
 
 class FaultComponent extends PureComponent {