+++ packages/react-renderer/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.8.12"></a>
+## [0.8.12](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-react-renderer@0.8.11...@ali/lowcode-react-renderer@0.8.12) (2020-06-24)
+
+
+### Bug Fixes
+
+* div 不显示问题 ([1b6533c](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/commit/1b6533c))
+
+
+
+
+<a name="0.8.11"></a>
+## [0.8.11](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-react-renderer@0.8.9...@ali/lowcode-react-renderer@0.8.11) (2020-06-23)
+
+
+
+
+**Note:** Version bump only for package @ali/lowcode-react-renderer
+
+<a name="0.8.9"></a>
+## [0.8.9](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-react-renderer@0.8.8...@ali/lowcode-react-renderer@0.8.9) (2020-06-23)
+
+
+### Bug Fixes
+
+* 修复低代码组件设计器、区块设计器根节点为 Page 的问题，修复 topArea 样式 ([e85b542](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/commit/e85b542))
+
+
+
+
 <a name="0.8.8"></a>
 ## [0.8.8](https://gitlab.alibaba-inc.com/ali-lowcode/ali-lowcode-engine/compare/@ali/lowcode-react-renderer@0.8.7...@ali/lowcode-react-renderer@0.8.8) (2020-05-15)